More believable clicks with relax mod
All in all, this is that the update does: whenever there is jumps, it will only use the left key, while if there is a stream of circles, it would start to switch keys. Just like how normal osu players play the game. The code I've written has comments so it should be easy to understand. I've written my code on these lines: 
269--273, 
280--283, 
304--315, 
380--381